pyproject.toml
[project]
name = "rl-model"
version = "0.1.0"
description = "Reinforcement Learning Model Service"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "transformers==4.37.0",
    "torch==2.1.2",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.0",
    "pydantic==2.5.3",
    "accelerate==0.26.1",
    "numpy<2", # Pin to 1.x for compatibility with compiled modules
    "huggingface-hub>=0.36.0",
    "safetensors>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
